+++ monolib-dart/lib/stream.dart
@@ -7,3 +7,4 @@ export 'src/stream/on_stream_of_lists.dart';
 export 'src/stream/single_value_caching_stream.dart';
 export 'src/stream/stream_with_close.dart';
 export 'src/stream/stream_where_type_extension.dart';
+export 'src/stream/map_not_null_stream_extension.dart';