+++ monolib-dart/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.23
 
 - Adds `StreamWhereTypeExtension` with `whereType<R>()` method to `Stream`.
+- Adds `MapNotNullStreamExtension` with `mapNotNull<R>()` method to `Stream`.
 
 ## 0.0.22